requirements.txt
sodapy==2.1.0
Flask==1.1.2
gunicorn==20.0.4
kaggle==1.5.6
newsapi-python==0.2.6
tweepy==3.9.0
requests==2.24.0
requests-oauthlib==1.3.0
json5==0.9.5
jsonschema==3.2.0
